requirements.txt
ddddocr
Flask
Flask-Limiter
flasgger
requests
urllib3==1.25.11
numpy==1.26.4
opencv-python
Pillow